requirements.txt
flask
pandas
google-cloud-bigquery
google-cloud-bigquery-storage
pyarrow
db-dtypes
gunicorn
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.1
Flask-Testing==0.8.1
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
packaging==23.2
pluggy==1.4.0
pytest==8.0.2
requests==2.31.0
urllib3==2.1.0
Werkzeug==3.0.1